sna: Check link-status after a hotplug event
If the modeset fails, or the link subsequently fails, we need to perform
the modeset again. To signal this the kernel sends us a hotplug uevent
with a new link-status property set to bad. The kernel may have to take
some corrective action which invalidates the current mode and so the
following modeset may fail and we need to go and report to the client
for them to choose the next course of action (reconfigure the displays).
At the very least the kernel *requires* us to reapply the current mode.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Signed-off-by: Manasi Navare <manasi.d.navare@intel.com>